+++ esp-hal-common/src/timer.rs
@@ -322,7 +322,7 @@ impl<TG> Timer0<TG>
 where
     TG: TimerGroupInstance,
 {
-    #[cfg(feature = "embassy-time-timg0")]
+    #[allow(unused)]
     pub(crate) unsafe fn steal() -> Self {
         Self {
             phantom: PhantomData,
@@ -483,7 +483,7 @@ impl<TG> Timer1<TG>
 where
     TG: TimerGroupInstance,
 {
-    #[cfg(feature = "embassy-time-timg1")]
+    #[allow(unused)]
     pub(crate) unsafe fn steal() -> Self {
         Self {
             phantom: PhantomData,